requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
chardet==3.0.4
cryptography
dj-database-url==0.5.0
Django==3.2.7
django-cors-headers==3.1.1
django-filter==2.4.0
django-heroku==0.3.1
djangorestframework==3.12.4
drf-jwt==1.13.4
gunicorn==20.1.0
idna==2.8
Markdown==3.3.4
psycopg2==2.9.1
pycparser==2.20
PyJWT==1.7.1
pytz==2021.1
requests==2.22.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.25.11
whitenoise==5.3.0
python-dotenv==0.19.0
django-mptt==0.13.4